+++ src/Exception/CouldNotRenderExceptionInterface.php
@@ -1,9 +1,11 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Dhii\Output\Exception;
 
 /**
- * Represents an exception thrown when the rendererer fails to render.
+ * Represents an exception thrown when the renderer fails to render.
  *
  * @since 0.1
  */