+++ src/RendererInterface.php
@@ -1,10 +1,13 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Dhii\Output;
 
-use Dhii\Util\String\StringableInterface as Stringable;
 use Dhii\Output\Exception\CouldNotRenderExceptionInterface;
 use Dhii\Output\Exception\RendererExceptionInterface;
+use Exception;
+use Stringable;
 
 /**
  * Something that can produce output.
@@ -20,6 +23,7 @@ interface RendererInterface
      *
      * @throws CouldNotRenderExceptionInterface If cannot render.
      * @throws RendererExceptionInterface       Any other problem related to the renderer.
+     * @throws Exception                        Any other problem.
      *
      * @return string|Stringable The output.
      */