+++ tests/unit/RendererInterfaceTest.php
@@ -1,0 +1,45 @@
+<?php
+
+namespace Dhii\Output\UnitTest;
+
+use Dhii\Output\RendererInterface as TestSubject;
+use PHPUnit\Framework\MockObject\MockObject as MockObject;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * Tests {@see TestSubject}.
+ *
+ * @since 0.1
+ */
+class RendererInterfaceTest extends TestCase
+{
+    /**
+     * Creates a new instance of the test subject.
+     *
+     * @since 0.1
+     *
+     * @return TestSubject|MockObject
+     */
+    public function createInstance()
+    {
+        $mock = $this->getMockBuilder(TestSubject::class)
+            ->getMock();
+
+        return $mock;
+    }
+
+    /**
+     * Tests whether a valid instance of the test subject can be created.
+     *
+     * @since 0.1
+     */
+    public function testCanBeCreated()
+    {
+        $subject = $this->createInstance();
+
+        $this->assertInstanceOf(
+            TestSubject::class, $subject,
+            'Could not create a valid instance of the test subject.'
+        );
+    }
+}